# docs/api/openapi.yaml
info:
  description: 'Human-facing REST API for QuantDinger. Agent integrations use `/api/agent/v1`
    — see `docs/agent/agent-openapi.json`. Conventions: `docs/API_CONVENTIONS.md`.'
  contact:
    name: QuantDinger
    url: https://github.com/brokermr810/quantdinger
  license:
    name: See repository LICENSE
  title: QuantDinger Web API
  version: 1.0.0
components:
  securitySchemes:
    HumanJWT:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 'User session JWT from `POST /api/auth/login`. Send as `Authorization:
        Bearer <token>`.'
    AgentToken:
      type: http
      scheme: bearer
      bearerFormat: qd_agent_xxx
      description: Agent Gateway token for `/api/agent/v1/*` only. See docs/agent/AGENT_QUICKSTART.md.
  schemas:
    Error:
      type: object
      properties:
        code:
          type: integer
          description: Error code
        status:
          type: string
          description: Error name
        message:
          type: string
          description: Error message
        errors:
          type: object
          description: Errors
          additionalProperties: {}
      additionalProperties: false
...
    HumanSuccessEnvelope:
      type: object
      properties:
        code:
          type: integer
          description: 1 on success
          example: 1
        msg:
          type: string
          example: success
        data:
          nullable: true
      additionalProperties: false
    HumanErrorEnvelope:
      type: object
      properties:
        code:
          type: integer
          description: 0 on error
          example: 0
        msg:
          type: string
        data:
          nullable: true
      additionalProperties: false
...
    CredentialCreateRequest:
      type: object
      properties:
        name:
          type: string
          default: ''
          maxLength: 128
        exchange_id:
          type: string
          minLength: 1
          maxLength: 32
        api_key:
          type: string
          default: ''
          maxLength: 4096
        apiKey:
          type: string
          default: ''
          maxLength: 4096
        secret_key:
          type: string
          default: ''
          maxLength: 4096
        secretKey:
          type: string
          default: ''
          maxLength: 4096
        passphrase:
          type: string
          default: ''
          maxLength: 4096
        base_url:
          type: string
          default: ''
          maxLength: 2048
          nullable: true
        baseUrl:
          type: string
          default: ''
          maxLength: 2048
          nullable: true
        enable_demo_trading:
          type: boolean
          default: false
        demo:
          type: boolean
          default: false
        sandbox:
          type: boolean
          default: false
        testnet:
          type: boolean
          default: false
        ibkr_host:
          type: string
          default: 127.0.0.1
          maxLength: 255
        ibkr_port:
          type: integer
          default: 7497
          minimum: 1
          maximum: 65535
        ibkr_client_id:
          type: integer
          default: 7
          minimum: 0
          maximum: 2147483647
        ibkr_account:
          type: string
          default: ''
          maxLength: 128
      required:
      - exchange_id
      additionalProperties: false
    CredentialCreatedData:
      type: object
      properties:
        id:
          type: integer
      required:
      - id
      additionalProperties: false
    CredentialCreatedResponse:
      type: object
      properties:
        code:
          type: integer
        msg:
          type: string
        data:
          anyOf:
          - type: object
            nullable: true
          - $ref: '#/components/schemas/CredentialCreatedData'
      required:
      - code
      - msg
      additionalProperties: false
...
  responses:
...
    UNPROCESSABLE_ENTITY:
      description: Unprocessable Entity
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
servers:
- url: http://localhost:5000
  description: Backend direct (python run.py)
- url: http://localhost:8888
  description: Docker Compose (via reverse proxy)
paths:
...
  /api/credentials/desktop-brokers-policy:
    get:
      responses:
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HumanErrorEnvelope'
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HumanSuccessEnvelope'
      summary: Whether IBKR (local TWS or IB Gateway) may be configured on this deployment.
      tags:
      - Credentials
      operationId: getApiCredentialsDesktopBrokersPolicy
      description: 'Whether IBKR (local TWS or IB Gateway) may be configured on this
        deployment.

        Frontend uses this to disable options and show guidance before save/test.'
      x-visibility: private
  /api/credentials/list:
    get:
      responses:
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HumanErrorEnvelope'
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HumanSuccessEnvelope'
      summary: List all credentials for the current user.
      tags:
      - Credentials
      operationId: getApiCredentialsList
      x-visibility: private
  /api/credentials/egress-ip:
    get:
      responses:
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HumanErrorEnvelope'
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HumanSuccessEnvelope'
      summary: Public egress IPv4/IPv6 of this API server (for exchange API key IP
        whitelist).
      tags:
      - Credentials
      operationId: getApiCredentialsEgressIp
      description: 'Public egress IPv4/IPv6 of this API server (for exchange API key
        IP whitelist).

        Uses ipify''s v4-only / v6-only endpoints so each family is detected independently.'
      x-visibility: private
  /api/credentials/create:
    post:
      responses:
        '422':
          $ref: '#/components/responses/UNPROCESSABLE_ENTITY'
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CredentialCreatedResponse'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HumanErrorEnvelope'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CredentialCreateRequest'
      summary: Create a new credential for the current user.
      description: Supports crypto exchanges, IBKR (US stocks), and Alpaca.
      tags:
      - Credentials
      operationId: postApiCredentialsCreate
      x-visibility: private
  /api/credentials/update-name:
    put:
      responses:
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HumanErrorEnvelope'
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HumanSuccessEnvelope'
      summary: Update display name (alias) only — API keys in encrypted_config are
        untouched.
      tags:
      - Credentials
      operationId: putApiCredentialsUpdateName
      x-visibility: private
    patch:
      responses:
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HumanErrorEnvelope'
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HumanSuccessEnvelope'
      summary: Update display name (alias) only — API keys in encrypted_config are
        untouched.
      tags:
      - Credentials
      operationId: patchApiCredentialsUpdateName
      x-visibility: private
  /api/credentials/delete:
    delete:
      parameters:
      - in: query
        name: id
        schema:
          type: integer
          minimum: 1
        required: true
      responses:
        '422':
          $ref: '#/components/responses/UNPROCESSABLE_ENTITY'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HumanErrorEnvelope'
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HumanSuccessEnvelope'
      summary: Delete a credential for the current user.
      tags:
      - Credentials
      operationId: deleteApiCredentialsDelete
      x-visibility: private
...
openapi: 3.0.3